Partial commit to Interval Tree implementation

+++ src/analyzers/sd3_analyzer/unit_tests/StrideTest.java
@@ -44,16 +44,16 @@ public class StrideTest {
     @Test
     void strideCreation() {
         logger.info("Initialising a non-empty test stride");
-        testStride1 = new Stride(BigInteger.valueOf(2000), BigInteger.valueOf(-4), BigInteger.valueOf(10), 11L, new PCPair(BigInteger.valueOf(123456), MemoryAccess.WRITE));
+        testStride1 = new Stride(BigInteger.valueOf(2000), BigInteger.valueOf(-4), BigInteger.valueOf(10), BigInteger.valueOf(11),  new PCPair(BigInteger.valueOf(123456), MemoryAccess.WRITE));
         testStride1.printStrideState();
 
-        Stride testStride2 = new Stride(BigInteger.valueOf(2000), BigInteger.valueOf(-4), BigInteger.valueOf(10), 11L, new PCPair(BigInteger.valueOf(123456), MemoryAccess.WRITE));
+        Stride testStride2 = new Stride(BigInteger.valueOf(2000), BigInteger.valueOf(-4), BigInteger.valueOf(10), BigInteger.valueOf(11),  new PCPair(BigInteger.valueOf(123456), MemoryAccess.WRITE));
         testStride2.printTestStrideState();
     }
 
     @TestFactory
     Stream<DynamicTest> testWithinBlock() {
-        testStride1 = new Stride(BigInteger.valueOf(1000), BigInteger.valueOf(4), BigInteger.valueOf(TEST_LIMIT), 15L, new PCPair(BigInteger.valueOf(123456), MemoryAccess.WRITE));
+        testStride1 = new Stride(BigInteger.valueOf(1000), BigInteger.valueOf(4), BigInteger.valueOf(TEST_LIMIT), BigInteger.valueOf(15),  new PCPair(BigInteger.valueOf(123456), MemoryAccess.WRITE));
 
         return accessArray1.stream().map(entry -> dynamicTest("For address " + InstructionsFileReader.toHexString(entry), () -> {
             Assertions.assertTrue(testStride1.addressWithinBlock(entry));
@@ -62,7 +62,7 @@ public class StrideTest {
 
     @TestFactory
     Stream<DynamicTest> testContainsAddressInStride() {
-        testStride1 = new Stride(BigInteger.valueOf(1000), BigInteger.valueOf(4), BigInteger.valueOf(TEST_LIMIT), 15L, new PCPair(BigInteger.valueOf(123456), MemoryAccess.WRITE));
+        testStride1 = new Stride(BigInteger.valueOf(1000), BigInteger.valueOf(4), BigInteger.valueOf(TEST_LIMIT), BigInteger.valueOf(15), new PCPair(BigInteger.valueOf(123456), MemoryAccess.WRITE));
 
         return accessArray2.stream().map(entry -> dynamicTest("For address " + entry, () -> {
             Assertions.assertTrue(testStride1.containsAddressInStride(entry));
@@ -74,7 +74,7 @@ public class StrideTest {
         logger.info("Test stride expansion version 1");
         logger.info("The initial state of the test stride is: " + testStride1.getStringStrideState());
         logger.info("Commence addition of addresses form an array stream.");
-        testStride1 = new Stride(BigInteger.valueOf(1000), BigInteger.valueOf(4), BigInteger.valueOf(1), 2L, new PCPair(BigInteger.valueOf(123456), MemoryAccess.WRITE));
+        testStride1 = new Stride(BigInteger.valueOf(1000), BigInteger.valueOf(4), BigInteger.valueOf(1), BigInteger.valueOf(2), new PCPair(BigInteger.valueOf(123456), MemoryAccess.WRITE));
 
         return accessArray1.stream().map(entry -> {
              testStride1.expandStride(entry);
@@ -90,7 +90,7 @@ public class StrideTest {
         logger.info("Test stride expansion version 2");
         logger.info("The initial state of the test stride is: " + testStride1.getStringStrideState());
         logger.info("Commence addition of addresses form an array stream.");
-        Stride testStride2 = new Stride(BigInteger.valueOf(1000), BigInteger.valueOf(4), BigInteger.valueOf(2), 2L, new PCPair(BigInteger.valueOf(123456), MemoryAccess.WRITE));
+        Stride testStride2 = new Stride(BigInteger.valueOf(1000), BigInteger.valueOf(4), BigInteger.valueOf(2), BigInteger.valueOf(2), new PCPair(BigInteger.valueOf(123456), MemoryAccess.WRITE));
         testStride2.printTestStrideState();
 
         return addressesToAdd1.stream().map(entry -> {
